Fix unicode strings conversion to str
We used to do a simple str() type conversion before to convert numbers to
strings when running scripts. However this didn't account for trying to run
script with unicode strings. So it could have happened that a script was
supposed to be run with unicode string which couldn't be converted into ascii
str. Instead we hardcode this conversion into utf-8 encoding

Fixes #9